+++ src/resolve.rs
@@ -49,6 +49,15 @@ impl MaybeResolve<Size<Option<f32>>> for Size<Dimension> {
     }
 }
 
+impl MaybeResolve<Size<Option<f32>>> for Size<Option<Dimension>> {
+    fn maybe_resolve(self, context: Size<Option<f32>>) -> Size<Option<f32>> {
+        Size {
+            width: self.width.and_then(|w| w.maybe_resolve(context.width)),
+            height: self.height.and_then(|h| h.maybe_resolve(context.height)),
+        }
+    }
+}
+
 impl ResolveOrDefault<Option<f32>, f32> for Dimension {
     /// Will return a default value of result is evaluated to `None`
     fn resolve_or_default(self, context: Option<f32>) -> f32 {
@@ -227,10 +236,10 @@ mod tests {
         ///
         /// The parent / context should not affect the outcome.
         #[rstest]
-        #[case(Size::AUTO, Size::NONE, Size::NONE)]
-        #[case(Size::AUTO, Size::new(5.0, 5.0), Size::NONE)]
-        #[case(Size::AUTO, Size::new(-5.0, -5.0), Size::NONE)]
-        #[case(Size::AUTO, Size::new(0.0, 0.0), Size::NONE)]
+        #[case(Size::<Dimension>::AUTO, Size::<Option<f32>>::NONE, Size::<Option<f32>>::NONE)]
+        #[case(Size::<Dimension>::AUTO, Size::new(5.0, 5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Dimension>::AUTO, Size::new(-5.0, -5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Dimension>::AUTO, Size::new(0.0, 0.0), Size::<Option<f32>>::NONE)]
         fn maybe_resolve_auto(
             #[case] input: Size<Dimension>,
             #[case] context: Size<Option<f32>>,
@@ -244,10 +253,10 @@ mod tests {
         ///
         /// The parent / context should not affect the outcome.
         #[rstest]
-        #[case(Size::from_points(5.0, 5.0), Size::NONE, Size::new(5.0, 5.0))]
-        #[case(Size::from_points(5.0, 5.0), Size::new(5.0, 5.0), Size::new(5.0, 5.0))]
-        #[case(Size::from_points(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(5.0, 5.0))]
-        #[case(Size::from_points(5.0, 5.0), Size::new(0.0, 0.0), Size::new(5.0, 5.0))]
+        #[case(Size::<Dimension>::from_points(5.0, 5.0), Size::<Option<f32>>::NONE, Size::new(5.0, 5.0))]
+        #[case(Size::<Dimension>::from_points(5.0, 5.0), Size::new(5.0, 5.0), Size::new(5.0, 5.0))]
+        #[case(Size::<Dimension>::from_points(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(5.0, 5.0))]
+        #[case(Size::<Dimension>::from_points(5.0, 5.0), Size::new(0.0, 0.0), Size::new(5.0, 5.0))]
         fn maybe_resolve_points(
             #[case] input: Size<Dimension>,
             #[case] context: Size<Option<f32>>,
@@ -262,12 +271,84 @@ mod tests {
         ///
         /// The context __should__ affect the outcome.
         #[rstest]
-        #[case(Size::from_percent(5.0, 5.0), Size::NONE, Size::NONE)]
-        #[case(Size::from_percent(5.0, 5.0), Size::new(5.0, 5.0), Size::new(25.0, 25.0))]
-        #[case(Size::from_percent(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(-25.0, -25.0))]
-        #[case(Size::from_percent(5.0, 5.0), Size::new(0.0, 0.0), Size::new(0.0, 0.0))]
+        #[case(Size::<Dimension>::from_percent(5.0, 5.0), Size::<Option<f32>>::NONE, Size::<Option<f32>>::NONE)]
+        #[case(Size::<Dimension>::from_percent(5.0, 5.0), Size::new(5.0, 5.0), Size::new(25.0, 25.0))]
+        #[case(Size::<Dimension>::from_percent(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(-25.0, -25.0))]
+        #[case(Size::<Dimension>::from_percent(5.0, 5.0), Size::new(0.0, 0.0), Size::new(0.0, 0.0))]
         fn maybe_resolve_percent(
             #[case] input: Size<Dimension>,
+            #[case] context: Size<Option<f32>>,
+            #[case] expected: Size<Option<f32>>,
+        ) {
+            assert_eq!(input.maybe_resolve(context), expected);
+        }
+    }
+
+    mod maybe_resolve_size_option_dimension {
+        use crate::{prelude::Size, resolve::MaybeResolve, style::Dimension};
+        use rstest::rstest;
+
+        /// Size<None> should always return Size<None>
+        ///
+        /// The parent / context should not affect the outcome.
+        #[rstest]
+        #[case(Size::<Option<Dimension>>::NONE, Size::<Option<f32>>::NONE, Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::NONE, Size::new(5.0, 5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::NONE, Size::new(-5.0, -5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::NONE, Size::new(0.0, 0.0), Size::<Option<f32>>::NONE)]
+        fn maybe_resolve_none(
+            #[case] input: Size<Option<Dimension>>,
+            #[case] context: Size<Option<f32>>,
+            #[case] expected: Size<Option<f32>>,
+        ) {
+            assert_eq!(input.maybe_resolve(context), expected);
+        }
+
+        /// Size<Dimension::Auto> should always return Size<None>
+        ///
+        /// The parent / context should not affect the outcome.
+        #[rstest]
+        #[case(Size::<Option<Dimension>>::AUTO, Size::<Option<f32>>::NONE, Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::AUTO, Size::new(5.0, 5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::AUTO, Size::new(-5.0, -5.0), Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::AUTO, Size::new(0.0, 0.0), Size::<Option<f32>>::NONE)]
+        fn maybe_resolve_auto(
+            #[case] input: Size<Option<Dimension>>,
+            #[case] context: Size<Option<f32>>,
+            #[case] expected: Size<Option<f32>>,
+        ) {
+            assert_eq!(input.maybe_resolve(context), expected);
+        }
+
+        /// Size<Dimension::Points> should always return a Size<Some(f32)>
+        /// where the f32 values are the inner value of the points.
+        ///
+        /// The parent / context should not affect the outcome.
+        #[rstest]
+        #[case(Size::<Option<Dimension>>::from_points(5.0, 5.0), Size::<Option<f32>>::NONE, Size::new(5.0, 5.0))]
+        #[case(Size::<Option<Dimension>>::from_points(5.0, 5.0), Size::new(5.0, 5.0), Size::new(5.0, 5.0))]
+        #[case(Size::<Option<Dimension>>::from_points(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(5.0, 5.0))]
+        #[case(Size::<Option<Dimension>>::from_points(5.0, 5.0), Size::new(0.0, 0.0), Size::new(5.0, 5.0))]
+        fn maybe_resolve_points(
+            #[case] input: Size<Option<Dimension>>,
+            #[case] context: Size<Option<f32>>,
+            #[case] expected: Size<Option<f32>>,
+        ) {
+            assert_eq!(input.maybe_resolve(context), expected);
+        }
+
+        /// `Size<Dimension::Percent>` should return `Size<None>` if context is `Size<None>`.
+        /// Otherwise it should return `Size<Some(f32)>`
+        /// where the f32 value is the inner value of the percent * context value.
+        ///
+        /// The context __should__ affect the outcome.
+        #[rstest]
+        #[case(Size::<Option<Dimension>>::from_percent(5.0, 5.0), Size::<Option<f32>>::NONE, Size::<Option<f32>>::NONE)]
+        #[case(Size::<Option<Dimension>>::from_percent(5.0, 5.0), Size::new(5.0, 5.0), Size::new(25.0, 25.0))]
+        #[case(Size::<Option<Dimension>>::from_percent(5.0, 5.0), Size::new(-5.0, -5.0), Size::new(-25.0, -25.0))]
+        #[case(Size::<Option<Dimension>>::from_percent(5.0, 5.0), Size::new(0.0, 0.0), Size::new(0.0, 0.0))]
+        fn maybe_resolve_percent(
+            #[case] input: Size<Option<Dimension>>,
             #[case] context: Size<Option<f32>>,
             #[case] expected: Size<Option<f32>>,
         ) {
@@ -370,7 +451,7 @@ mod tests {
         use rstest::rstest;
 
         #[rstest]
-        #[case(Rect::<Dimension>::AUTO, Size::NONE, Rect::ZERO)]
+        #[case(Rect::<Dimension>::AUTO, Size::<Option<f32>>::NONE, Rect::ZERO)]
         #[case(Rect::<Dimension>::AUTO, Size::new(5.0, 5.0), Rect::ZERO)]
         #[case(Rect::<Dimension>::AUTO, Size::new(-5.0, -5.0), Rect::ZERO)]
         #[case(Rect::<Dimension>::AUTO, Size::new(0.0, 0.0), Rect::ZERO)]
@@ -383,7 +464,7 @@ mod tests {
         }
 
         #[rstest]
-        #[case(Rect::<Dimension>::from_points(5.0, 5.0, 5.0, 5.0), Size::NONE, Rect::new(5.0, 5.0, 5.0, 5.0))]
+        #[case(Rect::<Dimension>::from_points(5.0, 5.0, 5.0, 5.0), Size::<Option<f32>>::NONE, Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Dimension>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(5.0, 5.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Dimension>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(-5.0, -5.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Dimension>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(0.0, 0.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
@@ -396,7 +477,7 @@ mod tests {
         }
 
         #[rstest]
-        #[case(Rect::<Dimension>::from_percent(5.0, 5.0, 5.0, 5.0), Size::NONE, Rect::ZERO)]
+        #[case(Rect::<Dimension>::from_percent(5.0, 5.0, 5.0, 5.0), Size::<Option<f32>>::NONE, Rect::ZERO)]
         #[case(Rect::<Dimension>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(5.0, 5.0), Rect::new(25.0, 25.0, 25.0, 25.0))]
         #[case(Rect::<Dimension>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(-5.0, -5.0), Rect::new(-25.0, -25.0, -25.0, -25.0))]
         #[case(Rect::<Dimension>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(0.0, 0.0), Rect::ZERO)]
@@ -416,7 +497,7 @@ mod tests {
         use rstest::rstest;
 
         #[rstest]
-        #[case(Rect::<Option<Dimension>>::NONE, Size::NONE, Rect::ZERO)]
+        #[case(Rect::<Option<Dimension>>::NONE, Size::<Option<f32>>::NONE, Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::NONE, Size::new(5.0, 5.0), Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::NONE, Size::new(-5.0, -5.0), Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::NONE, Size::new(0.0, 0.0), Rect::ZERO)]
@@ -429,7 +510,7 @@ mod tests {
         }
 
         #[rstest]
-        #[case(Rect::<Option<Dimension>>::AUTO, Size::NONE, Rect::ZERO)]
+        #[case(Rect::<Option<Dimension>>::AUTO, Size::<Option<f32>>::NONE, Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::AUTO, Size::new(5.0, 5.0), Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::AUTO, Size::new(-5.0, -5.0), Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::AUTO, Size::new(0.0, 0.0), Rect::ZERO)]
@@ -442,7 +523,7 @@ mod tests {
         }
 
         #[rstest]
-        #[case(Rect::<Option<Dimension>>::from_points(5.0, 5.0, 5.0, 5.0), Size::NONE, Rect::new(5.0, 5.0, 5.0, 5.0))]
+        #[case(Rect::<Option<Dimension>>::from_points(5.0, 5.0, 5.0, 5.0), Size::<Option<f32>>::NONE, Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Option<Dimension>>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(5.0, 5.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Option<Dimension>>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(-5.0, -5.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
         #[case(Rect::<Option<Dimension>>::from_points(5.0, 5.0, 5.0, 5.0), Size::new(0.0, 0.0), Rect::new(5.0, 5.0, 5.0, 5.0))]
@@ -455,7 +536,7 @@ mod tests {
         }
 
         #[rstest]
-        #[case(Rect::<Option<Dimension>>::from_percent(5.0, 5.0, 5.0, 5.0), Size::NONE, Rect::ZERO)]
+        #[case(Rect::<Option<Dimension>>::from_percent(5.0, 5.0, 5.0, 5.0), Size::<Option<f32>>::NONE, Rect::ZERO)]
         #[case(Rect::<Option<Dimension>>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(5.0, 5.0), Rect::new(25.0, 25.0, 25.0, 25.0))]
         #[case(Rect::<Option<Dimension>>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(-5.0, -5.0), Rect::new(-25.0, -25.0, -25.0, -25.0))]
         #[case(Rect::<Option<Dimension>>::from_percent(5.0, 5.0, 5.0, 5.0), Size::new(0.0, 0.0), Rect::ZERO)]