+++ RELEASES.md
@@ -6,11 +6,13 @@
 
 - Added `taffy::error::InvalidChild` Error type
 - `taffy::node::Taffy.new_leaf()` which allows the creation of new leaf-nodes without having to supply a measure function
+- Convenience constructors and const for `Size<Option<Dimension>>`
 
 ### 0.2.0 Changed
 
 - The `Default` of `Dimension` is now `Dimension::Auto` instead of the now removed `Dimension::Undefined`
 - `FlexboxLayout` fields `position`, `margin`, `padding` and `border` is now `Rect<Option<Dimension>>` instead of `Rect<Dimension>`
+- `FlexboxLayout` field `size` is now `Size<Option<Dimension>>` instead of `Size<Dimension>`
 - `Size<f32>.zero()` is now `Size::<f32>::ZERO`
 - `Point<f32>.zero()` is now  `Point::<f32>::ZERO`
 - renamed `taffy::node::Taffy.new_leaf()` -> `taffy::node::Taffy.new_leaf_with_measure()`